# swagger/ozonmp/ssn_service_api/v1/ssn_service_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ozonmp/ssn_service_api/v1/ssn_service_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "SsnServiceApiService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/services": {
      "get": {
        "summary": "ListServicesV1 - Services list",
        "operationId": "SsnServiceApiService_ListServicesV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListServicesV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "SsnServiceApiService"
        ]
      },
      "post": {
        "summary": "CreateServiceV1 - Create a service",
        "operationId": "SsnServiceApiService_CreateServiceV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateServiceV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateServiceV1Request"
            }
          }
        ],
        "tags": [
          "SsnServiceApiService"
        ]
      },
      "put": {
        "summary": "UpdateServiceV1 - Update a service",
        "operationId": "SsnServiceApiService_UpdateServiceV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateServiceV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateServiceV1Request"
            }
          }
        ],
        "tags": [
          "SsnServiceApiService"
        ]
      }
    },
    "/api/v1/services/{serviceId}": {
      "get": {
        "summary": "DescribeServiceV1 - Describe a service",
        "operationId": "SsnServiceApiService_DescribeServiceV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DescribeServiceV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "SsnServiceApiService"
        ]
      },
      "delete": {
        "summary": "RemoveServiceV1 - Remove a service",
        "operationId": "SsnServiceApiService_RemoveServiceV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RemoveServiceV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "SsnServiceApiService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateServiceV1Request": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "v1CreateServiceV1Response": {
      "type": "object",
      "properties": {
        "serviceId": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1DescribeServiceV1Response": {
      "type": "object",
      "properties": {
        "service": {
          "$ref": "#/definitions/v1Service"
        }
      }
    },
    "v1ListServicesV1Response": {
      "type": "object",
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Service"
          }
        }
      }
    },
    "v1RemoveServiceV1Response": {
      "type": "object",
      "properties": {
        "isFounded": {
          "type": "boolean"
        }
      }
    },
    "v1Service": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1UpdateServiceV1Request": {
      "type": "object",
      "properties": {
        "serviceId": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "v1UpdateServiceV1Response": {
      "type": "object"
    }
  }
}